# src/AdminAssistant.Infra.WebAPIClient/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Admin Assistant WebAPI v1.",
    "version": "v1"
  },
  "paths": {
    "/api/v1/accounts-module/BankAccount": {
      "put": {
        "tags": [
          "Accounts Module"
        ],
        "summary": "Update an existing BankAccount.",
        "operationId": "PutBankAccount",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BankAccountUpdateRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok - returns the updated BankAccountResponseDto",
            "schema": {
              "$ref": "#/definitions/BankAccountResponseDto"
            }
          },
          "404": {
            "description": "NotFound - When the BankAccountID of the given bankAccountUpdateRequest does not exist."
          },
          "422": {
            "description": "UnprocessableEntity - When the given bankAccountUpdateRequest is invalid."
          }
        }
      },
      "post": {
        "tags": [
          "Accounts Module"
        ],
        "summary": "Creates a new BankAccount.",
        "operationId": "PostBankAccount",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BankAccountCreateRequestDto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created - returns the created bank account with its assigned newly ID.",
            "schema": {
              "$ref": "#/definitions/BankAccountResponseDto"
            }
          },
          "422": {
            "description": "UnprocessableEntity - When the given bankAccountCreateRequest is invalid."
          }
        }
      }
    },
    "/api/v1/accounts-module/BankAccount/{bankAccountID}": {
      "get": {
        "tags": [
          "Accounts Module"
        ],
        "summary": "Gets the BankAccountResponseDto with the given ID.",
        "operationId": "GetBankAccountById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bankAccountID",
            "description": "The ID of the BankAccount to be returned.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - returns the BankAccount requested.",
            "schema": {
              "$ref": "#/definitions/BankAccountResponseDto"
            }
          },
          "404": {
            "description": "NotFound - When the given BankAccountID does not exist."
          }
        }
      }
    },
    "/api/v1/accounts-module/BankAccount/{bankAccountID}/transactions": {
      "get": {
        "tags": [
          "Accounts Module"
        ],
        "summary": "Get the transactions since the last bank account statement for the BankAccount with the given ID.",
        "operationId": "GetBankAccountTransactionByBankAccountID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bankAccountID",
            "description": "The ID of the BankAccount.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - returns a list of BankAccountTransactionResponseDto.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BankAccountTransactionResponseDto"
              }
            }
          },
          "404": {
            "description": "NotFound - When the given BankAccountID does not exist."
          }
        }
      }
    },
...
    "/api/v1/accounts-module/Bank": {
      "put": {
        "tags": [
          "Accounts Module"
        ],
        "summary": "Update an existing Bank.",
        "operationId": "PutBank",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BankUpdateRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok - returns the updated BankResponseDto",
            "schema": {
              "$ref": "#/definitions/BankResponseDto"
            }
          },
          "404": {
            "description": "NotFound - When the BankID of the given bankUpdateRequest does not exist."
          },
          "422": {
            "description": "UnprocessableEntity - When the given bankUpdateRequest is invalid."
          }
        }
      },
      "post": {
        "tags": [
          "Accounts Module"
        ],
        "summary": "Creates a new Bank.",
        "operationId": "PostBank",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BankCreateRequestDto"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created - returns the created bank with its assigned newly ID.",
            "schema": {
              "$ref": "#/definitions/BankResponseDto"
            }
          },
          "422": {
            "description": "UnprocessableEntity - When the given bankCreateRequest is invalid."
          }
        }
      },
      "get": {
        "tags": [
          "Accounts Module"
        ],
        "summary": "Lists all banks.",
        "operationId": "GetBank",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok - returns a list of BankResponseDto",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BankResponseDto"
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts-module/Bank/{bankID}": {
      "get": {
        "tags": [
          "Accounts Module"
        ],
        "summary": "Gets the Bank with the given ID.",
        "operationId": "GetBankById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bankID",
            "description": "The ID of the Bank to be returned.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - returns the Bank requested.",
            "schema": {
              "$ref": "#/definitions/BankResponseDto"
            }
          },
          "404": {
            "description": "NotFound - When the given BankID does not exist."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BankAccountCreateRequestDto": {
      "required": [
        "AccountName",
        "BankAccountTypeID",
        "CurrencyID",
        "IsBudgeted",
        "OpenedOn",
        "OpeningBalance"
      ],
      "type": "object",
      "properties": {
        "bankAccountTypeID": {
          "format": "int32",
          "description": "The BankAccountType for this BankAccount.",
          "type": "integer"
        },
        "currencyID": {
          "format": "int32",
          "type": "integer"
        },
        "accountName": {
          "type": "string"
        },
        "balance": {
          "format": "int32",
          "type": "integer"
        },
        "isBudgeted": {
          "type": "boolean"
        },
        "openingBalance": {
          "format": "int32",
          "type": "integer"
        },
        "currentBalance": {
          "format": "int32",
          "type": "integer"
        },
        "openedOn": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
    "BankAccountResponseDto": {
      "type": "object",
      "properties": {
        "bankAccountID": {
          "format": "int32",
          "description": "The BankAccount identifier.",
          "type": "integer",
          "readOnly": true
        },
        "bankAccountTypeID": {
          "format": "int32",
          "description": "The BankAccountType for this BankAccount.",
          "type": "integer"
        },
        "currencyID": {
          "format": "int32",
          "type": "integer"
        },
        "accountName": {
          "type": "string"
        },
        "isBudgeted": {
          "type": "boolean"
        },
        "openingBalance": {
          "format": "int32",
          "type": "integer"
        },
        "currentBalance": {
          "format": "int32",
          "type": "integer"
        },
        "openedOn": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "BankAccountTransactionResponseDto": {
      "type": "object",
      "properties": {
        "bankAccountTransactionID": {
          "format": "int32",
          "type": "integer"
        },
        "bankAccountID": {
          "format": "int32",
          "type": "integer"
        },
        "bankAccountTransactionTypeID": {
          "format": "int32",
          "type": "integer"
        },
        "bankAccountStatementID": {
          "format": "int32",
          "type": "integer"
        },
        "bankAccountStatementNumber": {
          "format": "int32",
          "type": "integer"
        },
        "isReconciled": {
          "type": "boolean"
        },
        "payeeID": {
          "format": "int32",
          "type": "integer"
        },
        "payeeName": {
          "type": "string"
        },
        "currencyID": {
          "format": "int32",
          "type": "integer"
        },
        "symbol": {
          "type": "string"
        },
        "decimalFormat": {
          "type": "string"
        },
        "credit": {
          "format": "double",
          "type": "number"
        },
        "debit": {
          "format": "double",
          "type": "number"
        },
        "balance": {
          "format": "double",
          "type": "number"
        },
        "description": {
          "type": "string"
        },
        "notes": {
          "type": "string"
        },
        "transactionDate": {
          "type": "string"
        }
      }
    },
...
    "BankAccountUpdateRequestDto": {
      "required": [
        "AccountName",
        "BankAccountID",
        "BankAccountTypeID",
        "CurrencyID",
        "IsBudgeted",
        "OpenedOn",
        "OpeningBalance"
      ],
      "type": "object",
      "properties": {
        "bankAccountID": {
          "format": "int32",
          "description": "The BankAccount identifier.",
          "type": "integer"
        },
        "bankAccountTypeID": {
          "format": "int32",
          "description": "The BankAccountType for this BankAccount.",
          "type": "integer"
        },
        "currencyID": {
          "format": "int32",
          "type": "integer"
        },
        "accountName": {
          "type": "string"
        },
        "isBudgeted": {
          "type": "boolean"
        },
        "openingBalance": {
          "format": "int32",
          "type": "integer"
        },
        "currentBalance": {
          "format": "int32",
          "type": "integer"
        },
        "openedOn": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "BankCreateRequestDto": {
      "required": [
        "BankName"
      ],
      "type": "object",
      "properties": {
        "bankName": {
          "type": "string"
        }
      }
    },
    "BankResponseDto": {
      "type": "object",
      "properties": {
        "bankID": {
          "format": "int32",
          "type": "integer"
        },
        "bankName": {
          "type": "string"
        }
      }
    },
    "BankUpdateRequestDto": {
      "required": [
        "BankID",
        "BankName"
      ],
      "type": "object",
      "properties": {
        "bankID": {
          "format": "int32",
          "description": "The Bank identifier.",
          "type": "integer",
          "readOnly": true
        },
        "bankName": {
          "type": "string"
        }
      }
    },
...
  }
}